Update backend logic to meet frontend
