+++ backend/.env.example
@@ -1,0 +1,5 @@
+PORT = "PORT"
+SECRET = "SOME_SECRET"
+MONGODB_URL = "MONGODB_URL"
+SENDER_EMAIL = "SENDER_EMAIL"
+SENDER_PASSWORD = "SENDER_PASSWORD"
