+++ backend/src/interfaces/models/Campaign.ts
@@ -11,7 +11,7 @@ interface ICampaign extends Document {
   title: string;
   excerpt: string;
   story?: string;
-  featured_image?: string;
+  image: string;
   fundRaiser?: string;
   wallet?: WalletOptions[];
   target: number;
