+++ backend/src/models/Campaign.model.ts
@@ -8,9 +8,9 @@ const CampaignModel = new mongoose.Schema<ICampaign>(
   {
     title: { type: String, minLength: 5, required: true },
     excerpt: { type: String, maxLength: 100, required: true },
-    story: { type: String },
-    featured_image: { type: String },
-    fundRaiser: { type: String },
+    story: String,
+    image: { type: String, required: true },
+    fundRaiser: String,
     wallet: [
       {
         name: String,
