+++ backend/src/server.ts
@@ -20,10 +20,7 @@ app.use(
 );
 
 app.use(cors());
-app.use(
-  '/images/users',
-  express.static(path.join(__dirname, '../images/users')),
-);
+app.use('/images', express.static(path.join(__dirname, '../images')));
 app.use(express.json());
 
 app.use(express.urlencoded({ extended: true }));
